Architect: Modularize finder_chart into finder_chart package
