+++ apts/visualization/finder_chart/constants.py
@@ -1,0 +1,82 @@
+THEMES = {
+    "stargazer_dark": {
+        "sky_bg": "#0B0F19",
+        "sky_top": "#0B0F19",
+        "sky_bottom": "#1E293B",
+        "ground_top": "#142319",
+        "ground_bottom": "#070E09",
+        "horizon_line": "#34D399",
+        "text_main": "#F8FAFC",
+        "text_sub": "#94A3B8",
+        "star": "#FFFFFF",
+        "star_glow": "#38BDF8",
+        "target_primary": "#38BDF8",
+        "target_secondary": "#FACC15",
+        "moon_bright": "#F1F5F9",
+        "moon_dark": "#1E293B",
+        "separation_line": "#38BDF8",
+        "separation_bg": "#0F172A",
+    },
+    "stargazer_light": {
+        "sky_bg": "#F8FAFC",
+        "sky_top": "#BAE6FD",
+        "sky_bottom": "#F8FAFC",
+        "ground_top": "#A7F3D0",
+        "ground_bottom": "#D1FAE5",
+        "horizon_line": "#059669",
+        "text_main": "#0F172A",
+        "text_sub": "#475569",
+        "star": "#1E293B",
+        "star_glow": "#0284C7",
+        "target_primary": "#0284C7",
+        "target_secondary": "#D97706",
+        "moon_bright": "#475569",
+        "moon_dark": "#CBD5E1",
+        "separation_line": "#0284C7",
+        "separation_bg": "#E2E8F0",
+    },
+}
+
+SKY_GRADIENT_COLORS = {
+    "DAY": ("#38BDF8", "#7DD3FC"),
+    "CIVIL_TWILIGHT": ("#1E1B4B", "#F43F5E"),
+    "NAUTICAL_TWILIGHT": ("0F172A", "#4338CA"),
+    "ASTRONOMICAL_TWILIGHT": ("#0B0F19", "#1E293B"),
+    "NIGHT_MOONLIT": ("#0F172A", "#1E293B"),
+    "NIGHT_DARK": ("#0B0F19", "#111827"),
+}
+
+COMPASS_POINTS = [
+    (0.0, "N"),
+    (22.5, "NNE"),
+    (45.0, "NE"),
+    (67.5, "ENE"),
+    (90.0, "E"),
+    (112.5, "ESE"),
+    (135.0, "SE"),
+    (157.5, "SSE"),
+    (180.0, "S"),
+    (202.5, "SSW"),
+    (225.0, "SW"),
+    (247.5, "WSW"),
+    (270.0, "W"),
+    (292.5, "WNW"),
+    (315.0, "NW"),
+    (337.5, "NNW"),
+    (360.0, "N"),
+]
+
+LONG_EVENT_CATEGORIES = {
+    "CONJUNCTION",
+    "PLANET_ALIGNMENT",
+    "OPPOSITION",
+    "CELESTIAL_CONFIGURATION",
+    "MOON_PHASE",
+    "SUPERMOON",
+    "MOON_LIBRATION",
+    "GREATEST_ELONGATION",
+    "VENUS_GREAT_BRILLIANCY",
+    "EQUINOX_SOLSTICE",
+    "METEOR_SHOWER",
+    "CELESTIAL_EVENT",
+}
